fix(sim): PR #580 deferred items — rope_scaling extraction, MaxModelLen int64, tests, docs
Complete 7 deferred hardening items from issue #580 (PR #587 handoff):

1. Extract applyRopeScaling as a pure function with 26 table-driven test
   cases covering blacklist (su/longrope/llama3), mrope fall-through,
   gemma3 substring match (handles text_config pivot), yarn original base,
   overflow guards, NaN/Inf defense, degenerate inputs.

2. Change MaxModelLen from int to int64 for consistency with ProgressIndex,
   TotalKVBlocks, BlockSizeTokens. Updates 6 type sites, removes redundant
   int64() casts, adds int64() widening at EnqueueRequest comparison sites.

3. Add cluster-mode MaxModelLen drop test (BC-6): Guard 1a (input >= limit)
   and Guard 1b (input + budget > limit), INV-1 conservation, inFlightRequests
   drain, Metrics.Requests map cleanup.

4. Add chunked prefill + MaxModelLen interaction test (BC-7): verifies no
   spurious force-completion during multi-chunk prefill (TotalOutputTokens=49,
   LengthCappedRequests=0, TTFT recorded).

5. Add glossary entries for MaxModelLen and Oracle Knowledge Boundary (INV-9).

6. Refine rope_scaling documentation with explicit blacklist details.

7. Fix pre-existing gemma3 bug: ParseHFConfig's text_config pivot overwrites
   model_type from "gemma3" to "gemma3_text", making the exact-match check
   dead code. Changed to strings.Contains to match vLLM's substring semantics.

Related to #580. Discovered issues: #602, #603, #604, #605.

+++ sim/simulator.go
@@ -76,7 +76,7 @@ type Simulator struct {
 	batchFormation       BatchFormation
 	model                  string
 	gpu                    string
-	maxModelLen            int // max total sequence length (0 = unlimited)
+	maxModelLen            int64 // max total sequence length (0 = unlimited)
 	rng                    *PartitionedRNG // partitioned RNG for deterministic multi-subsystem simulation
 	priorityPolicy         PriorityPolicy
 	scheduler              InstanceScheduler
@@ -108,9 +108,9 @@ func NewSimulator(cfg SimConfig, kvStore KVStore, latencyModel LatencyModel) (*S
 		if cfg.BlockSizeTokens <= 0 {
 			return nil, fmt.Errorf("NewSimulator: BlockSizeTokens must be > 0 when MaxModelLen is set, got %d", cfg.BlockSizeTokens)
 		}
-		// Overflow-safe ceiling division: avoids int64 wraparound for large MaxModelLen values (R11).
-		blocksForMaxLen := int64(cfg.MaxModelLen) / cfg.BlockSizeTokens
-		if int64(cfg.MaxModelLen)%cfg.BlockSizeTokens != 0 {
+		// Ceiling division for MaxModelLen → block count (R11).
+		blocksForMaxLen := cfg.MaxModelLen / cfg.BlockSizeTokens
+		if cfg.MaxModelLen%cfg.BlockSizeTokens != 0 {
 			blocksForMaxLen++
 		}
 		if blocksForMaxLen > cfg.TotalKVBlocks {
@@ -274,7 +274,7 @@ func (sim *Simulator) EnqueueRequest(r *Request) {
 	if sim.maxModelLen > 0 {
 		// vLLM uses >= for the input check (serving.py:1542): input that fills
 		// the entire context window leaves no room for even one output token.
-		if len(r.InputTokens) >= sim.maxModelLen {
+		if int64(len(r.InputTokens)) >= sim.maxModelLen {
 			logrus.Warnf("dropping request %s: input length %d >= MaxModelLen %d (no room for output)",
 				r.ID, len(r.InputTokens), sim.maxModelLen)
 			sim.Metrics.DroppedUnservable++
@@ -283,7 +283,7 @@ func (sim *Simulator) EnqueueRequest(r *Request) {
 		}
 		if r.MaxOutputLen > 0 {
 			// Client declared a budget: check input + budget fits context window
-			totalSeqLen := len(r.InputTokens) + r.MaxOutputLen
+			totalSeqLen := int64(len(r.InputTokens)) + int64(r.MaxOutputLen)
 			if totalSeqLen > sim.maxModelLen {
 				logrus.Warnf("dropping request %s: total sequence length %d (input=%d + budget=%d) exceeds MaxModelLen %d",
 					r.ID, totalSeqLen, len(r.InputTokens), r.MaxOutputLen, sim.maxModelLen)
@@ -498,7 +498,7 @@ func (sim *Simulator) processCompletions(now, currStepAdvance int64) []*Request 
 
 			// Record completion metrics
 			sim.recordRequestCompletion(req)
-		} else if sim.maxModelLen > 0 && req.ProgressIndex >= int64(sim.maxModelLen) {
+		} else if sim.maxModelLen > 0 && req.ProgressIndex >= sim.maxModelLen {
 			// BC-5: Runtime length cap — defense-in-depth.
 			// Force-complete any request that has reached MaxModelLen.
 			// This should not fire under normal operation (enqueue guard prevents it),
